requirements.txt
asyncpraw==7.8.1
anime-api==1.0.2
beautifulsoup4==4.12.3
cunnypy==3.0.0
hikari-crescent==1.0.0
hikari-miru==4.1.1
mkdocs-material==9.5.28
pysaucenao==1.6.2
